+++ quisp/backends/Backends.h
@@ -1,10 +1,11 @@
 #pragma once
 #include "ErrorTracking/Backend.h"
-#include "IQuantumBackend.h"
-#include "IQubitId.h"
+#include "interfaces/IQuantumBackend.h"
+#include "interfaces/IQubitId.h"
 
 // the namespace for exposing the backend
 namespace quisp::backends {
+
 using abstract::IQuantumBackend;
 using abstract::IQubit;
 using abstract::IQubitId;
