+++ quisp/backends/interfaces/IRandomNumberGenerator.h
